zsg: pad with block size only for last zone's data
Previously, every zone has same size of the data to be written:

  size_t each = left / ZSG_ZONES;

In this case, every zone may have small chunk of padded bytes to make it
aligned with block size (e.g., 4KB).  This patch writes data from the
first zone as much as possible and make the last chunk to be aligned
with the block size.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>